requirements.txt
playwright>=1.40
python-dotenv>=1.0
requests>=2.31
tabulate>=0.9
fpdf2>=2.7
Pillow>=10.0
pytesseract>=0.3
opencv-python>=4.8
numpy>=1.24
pytest>=7.0

